Replaced some property calls with protected variables where able

+++ app/src/main/java/com/example/duckgame/GameThread.java
@@ -47,7 +47,7 @@ public class GameThread extends Thread {
                to pass for an update loop, until the number of drawing frames skipped exceeds a set value. */
             while (deltaTime >= TIME_PER_TICK && framesSkipped < MAX_FRAMESKIP){
                 // Update all active objects on the level
-                level.tick(deltaTime);
+                level.tick((double)deltaTime);
                 deltaTime -= TIME_PER_TICK;
                 framesSkipped++;
             }
